got out pgrep as it is not available on osx

+++ capture/stream/controller.php
@@ -7,25 +7,26 @@ include_once("../../config.php");
 
 $idletime = 150;
 
-$procfile = file_get_contents(BASE_FILE."capture/stream/logs/procinfo");
+$pid = 0;
+$running = false;
+if(file_exists(BASE_FILE."capture/stream/logs/procinfo")) {
+	$procfile = file_get_contents(BASE_FILE."capture/stream/logs/procinfo");
 
-$tmp = explode("|", $procfile);
-$pid = $tmp[0];
-$last = $tmp[1];
+	$tmp = explode("|", $procfile);
+	$pid = $tmp[0];
+	$last = $tmp[1];
 
-//echo "no updates in " . (time() - $last) . " seconds<br />";
+	// check if script with pid is still runnning
+	exec("ps -p ".$pid." | wc -l",$out);
+	$out = trim($out[0]);
+	if($out == 2) 
+		$running = true;
 
-exec("pgrep php", $proclist, $return); 							// apache2 if script is started by Apache
+	// check whether the process has been idle for too long
+	logit("controller.log","script called - pid:" . $pid . "  idle:" . (time() - $last));
+	if(!$running || ( $last < time() - $idletime)) {
 
-logit("controller.log","script called - pid:" . $pid . "  idle:" . (time() - $last));
-
-if(in_array($pid,$proclist)) {
-
-	//echo "process " . $pid . " running";
-
-	if($last < time() - $idletime) {
-
-		exec("kill " . $pid);
+		if($running) exec("kill " . $pid);
 
 		logit("controller.log","script was idle for more than " . $idletime . " seconds - killing and starting");
 
@@ -33,8 +34,8 @@ if(in_array($pid,$proclist)) {
 
 		passthru("php ".BASE_FILE."capture/stream/capture.php > /dev/null 2>&1 &");
 	}
-
-} else {
+}
+if(!$running) {
 
 	logit("controller.log","script was not running - starting");
